+++ hyperlink/__init__.py
@@ -1,9 +1,16 @@
 
-from ._url import URL, URLParseError, register_scheme, parse_host
+from ._url import (URL,
+                   parse,
+                   EncodedURL,
+                   DecodedURL,
+                   URLParseError,
+                   register_scheme)
 
 __all__ = [
     "URL",
+    "parse",
+    "EncodedURL",
+    "DecodedURL",
     "URLParseError",
     "register_scheme",
-    "parse_host"
 ]