+++ hyperlink/test/test_url.py
@@ -5,6 +5,7 @@
 
 from __future__ import unicode_literals
 
+import sys
 import socket
 
 from .common import HyperlinkTestCase
@@ -13,6 +14,8 @@ from .. import URL, URLParseError
 from .. import _url
 from .._url import inet_pton, SCHEME_PORT_MAP, parse_host
 
+
+PY2 = (sys.version_info[0] == 2)
 unicode = type(u'')
 
 
@@ -329,6 +332,13 @@ class TestURL(HyperlinkTestCase):
         childURL = URL(host=u"www.foo.com").child(u"c")
         self.assertTrue(childURL.rooted)
         self.assertEqual("http://www.foo.com/c", childURL.to_text())
+
+    def test_emptyChild(self):
+        """
+        L{URL.child} without any new segments returns the original L{URL}.
+        """
+        url = URL(host=u"www.foo.com")
+        self.assertEqual(url.child(), url)
 
     def test_sibling(self):
         """
@@ -743,10 +753,14 @@ class TestURL(HyperlinkTestCase):
     def test_queryIterable(self):
         """
         When a L{URL} is created with a C{query} argument, the C{query}
-        argument is converted into an N-tuple of 2-tuples.
+        argument is converted into an N-tuple of 2-tuples, sensibly
+        handling dictionaries.
         """
+        expected = (('alpha', 'beta'),)
         url = URL(query=[['alpha', 'beta']])
-        self.assertEqual(url.query, (('alpha', 'beta'),))
+        self.assertEqual(url.query, expected)
+        url = URL(query={'alpha': 'beta'})
+        self.assertEqual(url.query, expected)
 
     def test_pathIterable(self):
         """
@@ -1146,4 +1160,38 @@ class TestURL(HyperlinkTestCase):
         assert norm_delimited_url.to_text() == '/a%2Fb/cd%3F?k%3D=v%23#test'
 
         # test invalid percent encoding during normalize
-        assert URL(path=('', '%te%sts')).normalize().to_text() == '/%te%sts'
+        assert URL(path=('', '%te%sts')).normalize(percents=False).to_text() == '/%te%sts'
+        assert URL(path=('', '%te%sts')).normalize().to_text() == '/%25te%25sts'
+
+        percenty_url = URL(scheme='ftp', path=['%%%', '%a%b'], query=[('%', '%%')], fragment='%', userinfo='%:%')
+
+        assert percenty_url.to_text(with_password=True) == 'ftp://%:%@/%%%/%a%b?%=%%#%'
+        assert percenty_url.normalize().to_text(with_password=True) == 'ftp://%25:%25@/%25%25%25/%25a%25b?%25=%25%25#%25'
+
+    def test_str(self):
+        # see also issue #49
+        text = u'http://example.com/á/y%20a%20y/?b=%25'
+        url = URL.from_text(text)
+        assert unicode(url) == text
+        assert bytes(url) == b'http://example.com/%C3%A1/y%20a%20y/?b=%25'
+
+        if PY2:
+            assert isinstance(str(url), bytes)
+            assert isinstance(unicode(url), unicode)
+        else:
+            assert isinstance(str(url), unicode)
+            assert isinstance(bytes(url), bytes)
+
+    def test_idna_corners(self):
+        text = u'http://abé.com/'
+        url = URL.from_text(text)
+        assert url.to_iri().host == u'abé.com'
+        assert url.to_uri().host == u'xn--ab-cja.com'
+
+        url = URL.from_text("http://ドメイン.テスト.co.jp#test")
+        assert url.to_iri().host == u'ドメイン.テスト.co.jp'
+        assert url.to_uri().host == u'xn--eckwd4c7c.xn--zckzah.co.jp'
+
+        assert url.to_uri().get_decoded_url().host == u'ドメイン.テスト.co.jp'
+
+        assert URL.from_text('http://Example.com').to_uri().get_decoded_url().host == 'example.com'